tdl curves

+++ dect_ping/src/dect/dect_phy_shell.h
@@ -54,8 +54,9 @@ struct dect_phy_perf_params {
 	/* Gap between HARQ feedback TX end and re-starting a server RX. */
 	uint8_t server_harq_feedback_tx_rx_delay_subslot_count;
 
-	/* PDC number for output reporting (e.g., pdc,100) */
-	uint16_t pdc_number;
+	/* PDC number for output reporting (e.g., pdc,100 or 99.5)
+	 * Stored as integer * 100 to support decimals (e.g., 9950 = 99.50) */
+	uint32_t pdc_number;
 };
 
 enum dect_phy_rf_tool_mode {
